+++ client/src/Settings.json
@@ -1,0 +1,6 @@
+{
+    "displayTime": false,
+    "darkMode": false,
+    "filter": false,
+    "selectedRoute": "['10', '15', '18', '19', '20', 'LOOP', 'UPPER CAMPUS','LOOP OUT OF SERVICE AT BARN THEATER','OUT OF SERVICE/SORRY','SPECIAL',]"
+}
